+++ CHANGELOG.md
@@ -9,7 +9,7 @@ This project adheres to semantic versioning.
 - Add `reject` function to the `Record` module.
 - Add `last`, `init`, `slice`, `lookup`, `toUpper`, and `toLower` functions to the `String` module.
 - Add `isValid`, `rem`, and `mod` functions to the `Number` module.
-- Add `construct`, `memoize`, and `uncurry2` through `uncurry5` functions to the `Function` module.
+- Add `construct`, `memoize`, `curry2` through `curry5`, `curry2T` through `curry5T`, and `uncurry2` through `uncurry5` functions to the `Function` module.
 - Add `Milliseconds` newtype to the `Date` module, and update the `Date` and `Task` modules to universally utilise it.
 - Provide the input value to the fallback function of `guard`.